requirements.txt
redis
django-redis-cache
requests
xmltodict